chore: clippy-clean ferric-rpa and its full dependency tree
Whole-crate clippy sweep (production code → 0 warnings, lib + all-targets
build clean). Two classes of action:

Mechanical fixes (behavior-preserving, mostly cargo clippy --fix):
- needless_borrow: drop redundant & on rhf.density_r()/mos_r() (already
  return refs) in ferric-mp2 gradient/zvector, ferric-scf pairs.
- empty eprintln!("") → eprintln!() in mp2 gradient test diagnostics.
- map .values() instead of (_, v) destructure in ferric-core basis.
- relocate scs_mp2_gradient_analytical above its test module (pure move).
- redundant `mut` on a non-capturing closure in properties::hirshfeld_i.

Documented lint policy for numerical code (crate- or site-level allows
with rationale, because the flagged patterns are deliberate):
- needless_range_loop: tensor/MO/grid index loops read clearer with
  explicit indices (ferric-rpa, -dft, -mp2, -quadrature, -scf, one site
  in -integrals).
- excessive_precision: reference tables (minimax nodes/weights, Lebedev
  grids) transcribed at full source precision on purpose.
- identity_op: `stride*g + 0` keeps component index aligned with +1/+2.
- op_ref: `a + &b + &b.t()` must borrow b (reused on the same line).
- too_many_arguments / type_complexity: self-energy and gradient kernels
  intrinsically thread many quantities / return multi-array tuples;
  introduced UhfFockMod<'a> alias where it genuinely reads better.

No numerics touched. ferric-rpa + dependency tree: 0 clippy warnings.

+++ crates/ferric-rpa/examples/gw100_ip_subset.rs
@@ -156,7 +156,7 @@ fn run_case_diag(case: &Case) -> Option<(f64, f64, f64, f64, CationDiag)> {
     // Parse neutral and cation as separate Molecule instances.
     let neutral = Molecule::parse_xyz(case.xyz, 0, 1).ok()?;
     let cation_mult: usize = {
-        let n = neutral.nelec() as i32 - 1;
+        let n = neutral.nelec() - 1;
         if n % 2 == 1 { 2 } else { 3 }  // doublet if odd; triplet for even-1
     };
     // Actually cation has nelec - 1; if neutral was singlet (even), cation
